Moved controller code
